Create Exosome Research Products Market to Reach USD 20.64 billion, by 2030 at 35.6% CAGR: Says the Data Bridge Market Research
Data Bridge Market Research analyzes that the global exosome research products market is expected to reach the value of USD 2,232,028.31 thousand by 2030, at a CAGR of 35.6% during the forecast period. Product accounts for the largest type segment in the market due to growing use of exosome research products in diagnostics and therapeutics.